+++ src/commonMain/kotlin/com/fleshgrinder/Uuid.kt
@@ -1,0 +1,17 @@
+@file:Suppress("AMBIGUOUS_ACTUALS")
+
+package com.fleshgrinder
+
+public expect class Uuid
+
+/**
+ * @throws IllegalArgumentException
+ */
+public expect fun String.toUuid(): Uuid
+
+public fun String.toUuidOrNull(): Uuid? =
+    try {
+        toUuid()
+    } catch (_: IllegalArgumentException) {
+        null
+    }