+++ CHANGELOG.md
@@ -6,5 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+### Added
+- UUID string parsing and formatting
 
 [Unreleased]: https://github.com/Fleshgrinder/kotlin-uuid